ffmpeg_sys_next

Function av_expr_parse

pub unsafe extern "C" fn av_expr_parse(
    expr: *mut *mut AVExpr,
    s: *const c_char,
    const_names: *const *const c_char,
    func1_names: *const *const c_char,
    funcs1: *const Option<unsafe extern "C" fn(arg1: *mut c_void, arg2: f64) -> f64>,
    func2_names: *const *const c_char,
    funcs2: *const Option<unsafe extern "C" fn(arg1: *mut c_void, arg2: f64, arg3: f64) -> f64>,
    log_offset: c_int,
    log_ctx: *mut c_void,
) -> c_int

Parse an expression.

@param expr a pointer where is put an AVExpr containing the parsed value in case of successful parsing, or NULL otherwise. The pointed to AVExpr must be freed with av_expr_free() by the user when it is not needed anymore. @param s expression as a zero terminated string, for example “1+2^3+5*5+sin(2/3)” @param const_names NULL terminated array of zero terminated strings of constant identifiers, for example {“PI”, “E”, 0} @param func1_names NULL terminated array of zero terminated strings of funcs1 identifiers @param funcs1 NULL terminated array of function pointers for functions which take 1 argument @param func2_names NULL terminated array of zero terminated strings of funcs2 identifiers @param funcs2 NULL terminated array of function pointers for functions which take 2 arguments @param log_offset log level offset, can be used to silence error messages @param log_ctx parent logging context @return >= 0 in case of success, a negative value corresponding to an AVERROR code otherwise